Fix: Fixes small differences between json and level objects

+++ Levels/LevelItems/LevelEnemy.cs
@@ -4,7 +4,7 @@
     {
         public string type { get; set; }
         public int startingX { get; set; }
-        public int startingY { get; set; }
+        public float startingY { get; set; }
         // true is right, false is left
         public bool direction { get; set; }
     }
